pyproject.toml
[tool.poetry]
name = "mbex"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.70.0"
pyzmq = "^22.3.0"
uvicorn = "^0.15.0"
bcrypt = "^3.2.0"
requests = "^2.26.0"
pydantic = {extras = ["email"], version = "^1.8.2"}
PyJWT = "^2.3.0"
attrs = "^21.2.0"
aioredis = "^2.0.0"

[tool.poetry.dev-dependencies]
black = "^21.11b1"
flake8 = "^4.0.1"
pytest = "^6.2.5"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
